Save devs

+++ project-management/src/main/java/project/management/services/BugTaskService.java
@@ -32,6 +32,6 @@ public interface BugTaskService {
 
     List<BugTaskResponseDto> getBugTasksByUserId(Long userId);
 
-    int countBugTasks();
+    long countBugTasks();
 
 }
